+++ src/protected/application/themes/BaseV1/layouts/parts/panel-space.php
@@ -1,3 +1,4 @@
+
 <article class="objeto clearfix">
     <?php if(isset($entity->{'@files:avatar.avatarSmall'}) && $avatar = $entity->{'@files:avatar.avatarSmall'}): ?>
         <div class="thumb" style="background-image: url(<?php echo $avatar->url; ?>)"></div>
@@ -24,23 +25,23 @@
 	</div>
     <div class="entity-actions">
         <a class="btn btn-small btn-primary" href="<?php echo $entity->editUrl; ?>"><?php \MapasCulturais\i::_e("editar");?></a>
+        <?php if(!isset($only_edit_button)): ?>
+            <?php if($entity->status === \MapasCulturais\Entities\Space::STATUS_ENABLED): ?>
+                <a class="btn btn-small btn-danger" href="<?php echo $entity->deleteUrl; ?>"><?php \MapasCulturais\i::_e("excluir");?></a>
+                <a class="btn btn-small btn-success" href="<?php echo $entity->archiveUrl; ?>"><?php \MapasCulturais\i::_e("arquivar");?></a>
 
-        <?php if($entity->status === \MapasCulturais\Entities\Space::STATUS_ENABLED): ?>
+            <?php elseif ($entity->status === \MapasCulturais\Entities\Space::STATUS_DRAFT): ?>
+                <a class="btn btn-small btn-warning" href="<?php echo $entity->publishUrl; ?>"><?php \MapasCulturais\i::_e("publicar");?></a>
+                <a class="btn btn-small btn-danger" href="<?php echo $entity->deleteUrl; ?>"><?php \MapasCulturais\i::_e("excluir");?></a>
 
-            <a class="btn btn-small btn-danger" href="<?php echo $entity->deleteUrl; ?>"><?php \MapasCulturais\i::_e("excluir");?></a>
-            <a class="btn btn-small btn-success" href="<?php echo $entity->archiveUrl; ?>"><?php \MapasCulturais\i::_e("arquivar");?></a>
+            <?php elseif ($entity->status === \MapasCulturais\Entities\Space::STATUS_ARCHIVED): ?>
+                <a class="btn btn-small btn-success" href="<?php echo $entity->unarchiveUrl; ?>"><?php \MapasCulturais\i::_e("desarquivar");?></a>
 
-        <?php elseif ($entity->status === \MapasCulturais\Entities\Space::STATUS_DRAFT): ?>
-            <a class="btn btn-small btn-warning" href="<?php echo $entity->publishUrl; ?>"><?php \MapasCulturais\i::_e("publicar");?></a>
-            <a class="btn btn-small btn-danger" href="<?php echo $entity->deleteUrl; ?>"><?php \MapasCulturais\i::_e("excluir");?></a>
-            
-        <?php elseif ($entity->status === \MapasCulturais\Entities\Space::STATUS_ARCHIVED): ?>
-            <a class="btn btn-small btn-success" href="<?php echo $entity->unarchiveUrl; ?>"><?php \MapasCulturais\i::_e("desarquivar");?></a>
-            
-        <?php else: ?>
-            <a class="btn btn-small btn-success" href="<?php echo $entity->undeleteUrl; ?>"><?php \MapasCulturais\i::_e("recuperar");?></a>
-            <?php if($entity->permissionTo->destroy): ?>
-                <a class="btn btn-small btn-danger" href="<?php echo $entity->destroyUrl; ?>"><?php \MapasCulturais\i::_e("excluir definitivamente");?></a>
+            <?php else: ?>
+                <a class="btn btn-small btn-success" href="<?php echo $entity->undeleteUrl; ?>"><?php \MapasCulturais\i::_e("recuperar");?></a>
+                <?php if($entity->permissionTo->destroy): ?>
+                    <a class="btn btn-small btn-danger" href="<?php echo $entity->destroyUrl; ?>"><?php \MapasCulturais\i::_e("excluir definitivamente");?></a>
+                <?php endif; ?>
             <?php endif; ?>
         <?php endif; ?>
     </div>
